Changed default get_keypoint method to call get_ramanan_keypoint by default.
Added access to ramanan keypoints for google dataset.

Also changed get_ramanan method so that it returns all ramanan-detected faces when it was run on the whole image,
but when it reads from ramanan__face??.mat files, that correspond to running ramanan to a Picasa cut-out face, it does not include the possibly multiple ramanan faces detected in that cutout but a single one.